+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `now.json` indentation.
 - Handling line items multiple images.
 - Add a missing `break` statement to Stripe function `processedLineItems` `switch` statement.
+- Add some missing `Async/Await` keywords to Stripe module.
 
 ## [0.1.0] - 2020-04-27
 ### Added